Moved ulter csv statement parser to  a service
