+++ services/ulster-statement-csv-parser/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "virtwallet-ulster-csv-parser",
+  "version": "1.0.0",
+  "description": "Virtwallet Ulster Bank Statement Parser Service",
+  "repository": "https://github.com/silvamb/virtwallet-serverless.git",
+  "author": "Matheus Silva <crush.comp04@gmail.com>",
+  "license": "GPLv3",
+  "dependencies": {
+    "libs": "file:../../libs",
+    "csv-parser": "^2.3.2",
+    "moment": "^2.24.0"
+  },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "mocha": "^7.0.0",
+    "serverless-iam-roles-per-function": "^2.0.2"
+  },
+  "scripts": {
+    "test": "mocha ./test/*-test.js"
+  }
+}
