+++ services/ulster-statement-csv-parser/sls-functions.yml
@@ -1,0 +1,23 @@
+sqs-handler:
+  handler: src/ulsterStatementCsvParserLambda.handle
+  environment:
+    transaction_queue_url:
+      Fn::ImportValue: Virtwallet-${self:provider.stage}-TransactionQueueUrl
+  events:
+    - sqs:
+        arn:
+          Fn::ImportValue: Virtwallet-${self:provider.stage}-TransactionQueueArn
+
+  iamRoleStatementsName: virtwallet-ulster-csv-parser-${self:provider.stage}-iam-role
+
+  iamRoleStatements:
+    - Effect: Allow
+      Action: 
+        - sqs:SendMessage
+        - sqs:ReceiveMessage
+      Resource:
+        Fn::ImportValue: Virtwallet-${self:provider.stage}-TransactionQueueArn
+    - Effect: Allow
+      Action: 
+        - s3:GetObject
+      Resource: "arn:aws:s3:::${self:custom.userFilesBucket}/account-files*"
