+++ services/ulster-statement-csv-parser/test/ulsterStatementCsvParser-test.js
@@ -2,15 +2,15 @@ const chai = require("chai");
 const chaiAsPromised = require("chai-as-promised");
 chai.use(chaiAsPromised);
 const expect = chai.expect;
-const fs = require('fs')
+const fs = require('fs');
 
-const parser = require('../../src/lambdas/ulster-statement-csv-parser/ulsterStatementCsvParser');
+const parser = require('../src/ulsterStatementCsvParser');
 
 const CsvParser = parser.UlsterCsvParser;
 
 describe('UlsterStatementCsvParserTest', () => {
     it('should parse file with success', () => {
-        const fileStream = fs.createReadStream('./test/unit/csv/ACCOUNTNAME12345678-20200105.csv');
+        const fileStream = fs.createReadStream('./test/csv/ACCOUNTNAME12345678-20200105.csv');
         const expectedBucket = "someS3Bucket";
         const expectedKey = "s3Key";
